refactor(app): give the Pipeline tab a slice of its own
Structure review, finding 15 — the structural item to do first, since it is
the only one that creates a slice and it unblocks two misfilings.

Five of the six results tabs delegated to a feature or a shared panel;
`pipeline` alone was assembled from raw parts in the app shell — the rail, a
stage card built from three file-local components, and the rewrite stepper.
That is why `ResultsColumn.tsx` was 580 lines when its actual job (banners,
the panels record, the tab shell) is a fraction of it. It is 410 now.

The promotion rule had already diagnosed this mechanically, and the two
misfilings it named move with the tab:

  components/StageDiff.tsx          -> features/pipeline/   (one consumer, and
                                       it was this card)
  features/editor/StageLayerEditor  -> features/pipeline/   (its only consumer
                                       was this card — a cross-slice edge that
                                       survived only because the SHELL was
                                       doing the importing)

Verified rather than assumed before moving: `AdvancedZone` merely mentions
`StageLayerEditor` in a comment, it does not import it. `StageRail` and
`MigrationSteps` stay in components/ — each has a genuine second consumer
(`ConfigColumn`'s `StageRailPreview`, and `PresetDetail`).

`EmptyNote` was local to the shell; the pipeline needs the same note for its
no-rewrites case, so it moves to `components/` as a real two-consumer shared
component rather than being copied — which is how the two `formatSnippet`s
happened.

The `deferredStage` vs `selectedStage` distinction survives verbatim, and the
props interface now says WHY the two are not interchangeable: `selectedStage`
drives the inputs (an input lagging a click by a frame eats the first
keystroke), `deferredStage` drives the expensive diff and stepper. The seven
layer props stay seven props rather than becoming one object, because the
shell forwards them straight through and bundling them into a per-render value
is exactly how roadmap 032's "typing reconciles no panel" memo contract gets
lost.

Verified: 959 unit/component/shimmed tests green, and the full 127-test e2e
suite green — including `keystroke-render.shimmed.test.tsx`, which is what
covers the deferred/selected distinction.
